Push step creation into BaseWizardBatchBuilder
